requirements.txt
pychrome
pyyaml
click
watchdog
voluptuous
tornado
tornado-http-auth